==> hardwarelist/2_merge_tables.ipynb <==
import pandas as pd
import os
import gzip
import pickle
import openai
import re
import copy

old = pd.read_csv('hardware_bitcoinwiki.csv') #hardware_name,hashrate,efficiency,hardware_index
new = pd.read_csv('hardware_asicminervalue.csv') #hardware_name,hardware_release_date,speed,power,noise,hash,profit,Mhash/J,hardware_index

# rename columns
old = old.rename(columns={"hashrate": "speed", "efficiency": "Mhash/J"})

old = old[['hardware_name', "Mhash/J"]]
new = new[['hardware_name', "Mhash/J"]]

# merge the two hardware csvs
hardware_df = pd.concat([old, new]).reset_index(drop=True)

# remove rows with "unknown" Mhash/J
hardware_df = hardware_df[hardware_df['Mhash/J'] != "unknown"]

# make all names lowercase
hardware_df['hardware_name'] = hardware_df['hardware_name'].str.lower()

# remove duplicate hardware names
hardware_df = hardware_df.drop_duplicates(subset=['hardware_name'])

# sort by hardware_name
hardware_df = hardware_df.sort_values(by=['hardware_name'])


hardware_df

hardware_df.to_csv('hardware_merged.csv', index=False)
