==> hardwarelist/4_create_pmaxv3.ipynb <==
import pandas as pd
from datetime import datetime

pmaxv1 = pd.read_csv('pmaxv1/pmaxv1.csv') # rows are date,max (TH/J),useless,archaicity (TH/J)
pmaxv1

pmaxv1.index = pd.to_datetime(pmaxv1['date'])

pmaxv2 = pd.read_csv('../hardwarelist/pmaxv2.csv') # date	max_efficiency
pmaxv2['date'] = pd.to_datetime(pmaxv2['date'])
pmaxv2 = pmaxv2.set_index('date')
pmaxv2

# for each date in pmaxv2, get the max between the two dataframes
pmaxv3 = pd.DataFrame(columns=['date', 'max_efficiency'])
for date in pmaxv2.index:
    max1 = pmaxv1[pmaxv1.index <= date].iloc[-1]['max (TH/J)']*1e12
    max2 = pmaxv2.loc[date]['max_efficiency']
    pmaxv3 = pd.concat([pmaxv3, pd.DataFrame([[date, max(max1, max2)]], columns=['date', 'max_efficiency'])])

pmaxv3.to_csv('../hardwarelist/pmaxv3.csv', index=False)
